requirements.txt
gpxpy
geopandas
shapely
pyproj
pandas
numpy
rasterio
srtm.py
folium
requests
